Use the repo code

+++ dimos/blockchain.go
@@ -1,5 +1,5 @@
 package dimos
 
 type Blockchain struct {
-    Height Int
+    Height int64
 }
